refactored code to match v5.6.0 on the vonage repo
