+++ src/main/java/com/nexmo/client/application/DeleteApplicationMethod.java
@@ -21,6 +21,7 @@
  */
 package com.nexmo.client.application;
 
+import com.nexmo.client.AbstractMethod;
 import com.nexmo.client.HttpWrapper;
 import com.nexmo.client.NexmoBadRequestException;
 import com.nexmo.client.NexmoClientException;
@@ -32,7 +33,7 @@ import org.apache.http.util.EntityUtils;
 import java.io.IOException;
 import java.io.UnsupportedEncodingException;
 
-class DeleteApplicationMethod extends ApplicationMethod<String, Void> {
+class DeleteApplicationMethod extends AbstractMethod<String, Void> {
     private static final Class[] ALLOWED_AUTH_METHODS = new Class[]{TokenAuthMethod.class};
 
     private static final String PATH = "/applications/%s";
@@ -60,5 +61,10 @@ class DeleteApplicationMethod extends ApplicationMethod<String, Void> {
         }
 
         return null;
+    }
+
+    @Override
+    protected RequestBuilder applyAuth(RequestBuilder request) throws NexmoClientException {
+        return getAuthMethod(getAcceptableAuthMethods()).applyAsBasicAuth(request);
     }
 }
