+++ src/test/java/com/nexmo/client/application/AppBasicAuthTest.java
@@ -21,20 +21,24 @@
  */
 package com.nexmo.client.application;
 
-import com.nexmo.client.AbstractMethod;
 import com.nexmo.client.HttpWrapper;
-import com.nexmo.client.NexmoClientException;
+import com.nexmo.client.auth.TokenAuthMethod;
 import org.apache.http.client.methods.RequestBuilder;
+import org.junit.Test;
 
-public abstract class ApplicationMethod<RequestT, ResultT> extends AbstractMethod<RequestT, ResultT> {
+import static org.junit.Assert.assertEquals;
 
+public class AppBasicAuthTest {
 
-    public ApplicationMethod(HttpWrapper httpWrapper) {
-        super(httpWrapper);
-    }
+    @Test
+    public void testApplyAuth() throws Exception {
+        String expectedAuthStr = "Authorization: Basic NHQ4YWVhZ2I6WHlNczJKa21BMVp6OWRVaw==";
+        HttpWrapper wrapper = new HttpWrapper(new TokenAuthMethod("4t8aeagb", "XyMs2JkmA1Zz9dUk"));
+        CreateApplicationMethod method = new CreateApplicationMethod(wrapper);
+        RequestBuilder request = method.makeRequest(Application.builder().build());
+        RequestBuilder requestWithAuth = method.applyAuth(request);
 
-    @Override
-    protected RequestBuilder applyAuth(RequestBuilder request) throws NexmoClientException {
-        return getAuthMethod(getAcceptableAuthMethods()).applyAsBasicAuth(request);
+        assertEquals(expectedAuthStr, String.valueOf(requestWithAuth.getHeaders("Authorization")[0]));
+
     }
 }
