--- src/test/java/com/nexmo/client/application/ApplicationMethodTest.java
@@ -1,44 +1,0 @@
-/*
- * Copyright (c) 2011-2019 Nexmo Inc
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- * THE SOFTWARE.
- */
-package com.nexmo.client.application;
-
-import com.nexmo.client.HttpWrapper;
-import com.nexmo.client.auth.TokenAuthMethod;
-import org.apache.http.client.methods.RequestBuilder;
-import org.junit.Test;
-
-import static org.junit.Assert.assertEquals;
-
-public class ApplicationMethodTest {
-
-    @Test
-    public void testApplyAuth() throws Exception {
-        String expectedAuthStr = "Authorization: Basic NHQ4YWVhZ2I6WHlNczJKa21BMVp6OWRVaw==";
-        HttpWrapper wrapper = new HttpWrapper(new TokenAuthMethod("4t8aeagb", "XyMs2JkmA1Zz9dUk"));
-        CreateApplicationMethod method = new CreateApplicationMethod(wrapper);
-        RequestBuilder request = method.makeRequest(Application.builder().build());
-        RequestBuilder requestWithAuth = method.applyAuth(request);
-
-        assertEquals(expectedAuthStr, String.valueOf(requestWithAuth.getHeaders("Authorization")[0]));
-
-    }
-}
