+++ src/test/java/com/nexmo/client/application/GetApplicationMethodTest.java
@@ -29,7 +29,7 @@ import org.junit.Test;
 
 import static org.junit.Assert.assertEquals;
 
-public class GetApplicationMethodTest extends ApplicationMethodTest {
+public class GetApplicationMethodTest extends AppBasicAuthTest {
     private GetApplicationMethod method;
 
     @Before
