+++ src/test/java/com/nexmo/client/application/UpdateApplicationMethodTest.java
@@ -29,7 +29,7 @@ import org.junit.Test;
 
 import static org.junit.Assert.assertEquals;
 
-public class UpdateApplicationMethodTest extends ApplicationMethodTest {
+public class UpdateApplicationMethodTest extends AppBasicAuthTest {
     private UpdateApplicationMethod method;
 
     @Before
